reformat tests into their own files
